# Improve performance of Enum.{Try}Parse