… src/System.Private.CoreLib/src/System/Enum.cs