@stephentoub 2018-11-28
@dotnet-bot test CentOS7.1 x64 Debug Innerloop Build please
@dotnet-bot test Linux-musl x64 Debug Build please
@dotnet-bot test OSX10.12 x64 Checked Innerloop Build and Test please
@dotnet-bot test Tizen armel Cross Checked Innerloop Build and Test please