@stephentoub 2018-11-29
> Could you please fix the conflict?

Rebased / fixed conflicts.