@TylerBrinkley 2018-12-18
@stephentoub How did you get those performance results? I'd like to do the same with my changes at https://github.com/dotnet/coreclr/compare/master...TylerBrinkley:enum-perf.