@stephentoub 2018-12-18
Benchmark.NET, then copying/pasting the results into Excel to compare.  (With the --coreRun argument that was recently added to Benchmark.NET, this can be achieved more easily with the right set up locally.)