@TylerBrinkley 2019-01-09
@stephentoub alright, I got things setup locally to benchmark my changes with the --coreRun argument. Do you happen to have the tests you used for this? I'd like to benchmark them with my changes.