@TylerBrinkley 2019-01-10
Thanks, I saw that. I'll try to expand on those.